pyproject.toml
[project]
name = "sample-fastapi-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
    "pymongo>=3.12.0,<4.0.0",
    "python-dotenv>=1.0.0",
]
